+++ tests/testthat/test-covlmc_draw.R
@@ -1,21 +1,38 @@
 test_that("draw obeys its contract (with vgam)", {
   withr::local_seed(0)
   dts <- sample(c("A", "B", "C"), 1000, replace = TRUE)
-  y <- ifelse(runif(length(dts)) > 0.5, c(dts[-1], sample(c("A", "B", "C"), 1)), c(dts[-c(1, 2)], sample(c("A", "B", "C"), 2, replace = TRUE)))
-  y <- as.factor(ifelse(runif(length(dts)) > 0.2, y, sample(c("A", "B", "C"), length(dts), replace = TRUE)))
+  y <- ifelse(runif(length(dts)) > 0.5, c(dts[-1], sample(c("A", "B", "C"), 1)),
+    c(dts[-c(1, 2)], sample(c("A", "B", "C"), 2, replace = TRUE))
+  )
+  y <- as.factor(ifelse(runif(length(dts)) > 0.2, y,
+    sample(c("A", "B", "C"), length(dts), replace = TRUE)
+  ))
   df_y <- data.frame(y = y)
   model <- covlmc(dts, df_y, alpha = 0.01, min_size = 1.5)
   expect_snapshot_output(draw(model, model = NULL, p_value = TRUE))
   expect_snapshot_output(draw(prune(model, 0.0001)))
-  expect_snapshot_output(draw(prune(model, 0.0001), control = draw_control(time_sep = " % ")))
-  expect_snapshot_output(draw(prune(model, 0.0001), model = "full", control = draw_control(time_sep = " % ")))
-  expect_snapshot_output(draw(prune(model, 0.0001), model = "full", control = draw_control(time_sep = " % "), with_state = TRUE))
+  expect_snapshot_output(draw(prune(model, 0.0001),
+    control = draw_control(time_sep = " % ")
+  ))
+  expect_snapshot_output(draw(prune(model, 0.0001),
+    model = "full",
+    control = draw_control(time_sep = " % ")
+  ))
+  expect_snapshot_output(draw(prune(model, 0.0001),
+    model = "full",
+    control = draw_control(time_sep = " % "),
+    with_state = TRUE
+  ))
   expect_snapshot_output(draw(prune(model, 0.0001),
     model = "full",
     control = draw_control(time_sep = " % ", intercept_sep = " + "),
     with_state = TRUE
   ))
-  expect_snapshot_output(draw(prune(model, 0.0001), model = "coef", control = draw_control(time_sep = " % "), with_state = TRUE))
+  expect_snapshot_output(draw(prune(model, 0.0001),
+    model = "coef",
+    control = draw_control(time_sep = " % "),
+    with_state = TRUE
+  ))
   expect_snapshot_output(draw(prune(model, 0.0001),
     model = "coef", control = draw_control(time_sep = " % ", level_sep = " @ "),
     with_state = TRUE,
@@ -24,33 +41,54 @@ test_that("draw obeys its contract (with vgam)", {
   ## skip_on_ci()
   ##  expect_snapshot_output(draw(model))
   ##  expect_snapshot_output(draw(model, control = draw_control(time_sep = " % ")))
-  ##  expect_snapshot_output(draw(model, digits = 3))
-  ##  expect_snapshot_output(draw(model, model = NULL, digits = 2))
-  ##  expect_snapshot_output(draw(model, p_value = FALSE, digits = 1))
-  ##  expect_snapshot_output(draw(model, model = "full", time_sep = " ~ ", digits = 1))
-  ##  expect_snapshot_output(draw(model, model = "full", time_sep = " ~ ", digits = 5, with_state = TRUE))
-  ##  expect_snapshot_output(draw(model, model = "coef", time_sep = " ~ ", digits = 5, with_state = TRUE))
+  ##  expect_snapshot_output(draw(model, control = draw_control(digits = 3)))
+  ##  expect_snapshot_output(draw(model, model = NULL, control = draw_control(digits = 2)))
+  ##  expect_snapshot_output(draw(model, p_value = FALSE, control = draw_control(digits = 1)))
+  ##  expect_snapshot_output(draw(model, model = "full", time_sep = " ~ ", control = draw_control(digits = 1)))
+  ##  expect_snapshot_output(draw(model, model = "full", time_sep = " ~ ", control = draw_control(digits = 5), with_state = TRUE))
+  ##  expect_snapshot_output(draw(model, model = "coef", time_sep = " ~ ", control = draw_control(digits = 5), with_state = TRUE))
 })
 
 test_that("draw obeys its contract (with nnet)", {
   withr::local_seed(0)
   withr::local_options(mixvlmc.predictive = "multinom")
   dts <- sample(c("A", "B", "C"), 1000, replace = TRUE)
-  y <- ifelse(runif(length(dts)) > 0.5, c(dts[-1], sample(c("A", "B", "C"), 1)), c(dts[-c(1, 2)], sample(c("A", "B", "C"), 2, replace = TRUE)))
-  y <- as.factor(ifelse(runif(length(dts)) > 0.2, y, sample(c("A", "B", "C"), length(dts), replace = TRUE)))
+  y <- ifelse(runif(length(dts)) > 0.5, c(dts[-1], sample(c("A", "B", "C"), 1)),
+    c(dts[-c(1, 2)], sample(c("A", "B", "C"), 2, replace = TRUE))
+  )
+  y <- as.factor(ifelse(runif(length(dts)) > 0.2, y,
+    sample(c("A", "B", "C"), length(dts), replace = TRUE)
+  ))
   df_y <- data.frame(y = y)
   model <- covlmc(dts, df_y, alpha = 0.01, min_size = 1.5)
   expect_snapshot_output(draw(model))
   expect_snapshot_output(draw(model, control = draw_control(time_sep = " % ")))
-  expect_snapshot_output(draw(model, digits = 3))
-  expect_snapshot_output(draw(model, model = NULL, digits = 2))
-  expect_snapshot_output(draw(model, p_value = FALSE, digits = 1))
-  expect_snapshot_output(draw(model, model = "full", time_sep = " ^ ", digits = 1))
-  expect_snapshot_output(draw(model, model = "full", time_sep = " ^ ", digits = 3, with_state = TRUE))
-  expect_snapshot_output(draw(model, model = "coef", time_sep = " ^ ", digits = 3, with_state = TRUE))
+  expect_snapshot_output(draw(model, control = draw_control(digits = 3)))
   expect_snapshot_output(draw(model,
-    model = "coef", time_sep = " ^ ", digits = 3,
-    with_state = TRUE, control = draw_control(level_sep = " @ ")
+    model = NULL,
+    control = draw_control(digits = 2)
+  ))
+  expect_snapshot_output(draw(model,
+    p_value = FALSE,
+    control = draw_control(digits = 1)
+  ))
+  expect_snapshot_output(draw(model,
+    model = "full", time_sep = " ^ ",
+    control = draw_control(digits = 1)
+  ))
+  expect_snapshot_output(draw(model,
+    model = "full", time_sep = " ^ ",
+    control = draw_control(digits = 3),
+    with_state = TRUE
+  ))
+  expect_snapshot_output(draw(model,
+    model = "coef", time_sep = " ^ ",
+    control = draw_control(digits = 3),
+    with_state = TRUE
+  ))
+  expect_snapshot_output(draw(model,
+    model = "coef", time_sep = " ^ ",
+    with_state = TRUE, control = draw_control(digits = 3, level_sep = " @ ")
   ))
 })
 
@@ -69,11 +107,14 @@ test_that("draw handles cases when levels have been dropped", {
   z <- runif(length(x)) + c(x[-1], 0) / 4
   dts_cov <- data.frame(y = y, z = z)
   m_cov <- covlmc(x = x, covariate = dts_cov, min_size = 3, alpha = 0.5)
-  expect_snapshot_output(draw(m_cov, model = "full", control = draw_control(time_sep = " % "), digits = 1))
   expect_snapshot_output(draw(m_cov,
     model = "full",
-    control = draw_control(time_sep = " % ", intercept_sep = " + "),
-    digits = 2, with_state = TRUE
+    control = draw_control(time_sep = " % ", digits = 1)
+  ))
+  expect_snapshot_output(draw(m_cov,
+    model = "full",
+    control = draw_control(time_sep = " % ", intercept_sep = " + ", digits = 2),
+    with_state = TRUE
   ))
 })
 
@@ -93,11 +134,14 @@ test_that("draw handles cases when multinom is used for two states time series",
   z <- runif(length(x)) + c(x[-1], 0) / 4
   dts_cov <- data.frame(y = y, z = z)
   m_cov <- covlmc(x = x, covariate = dts_cov, min_size = 3, alpha = 0.5)
-  expect_snapshot_output(draw(m_cov, model = "full", control = draw_control(time_sep = " % "), digits = 1))
   expect_snapshot_output(draw(m_cov,
     model = "full",
-    control = draw_control(time_sep = " % ", intercept_sep = " + "),
-    digits = 1, with_state = TRUE
+    control = draw_control(time_sep = " % ", digits = 1)
+  ))
+  expect_snapshot_output(draw(m_cov,
+    model = "full",
+    control = draw_control(digits = 1, time_sep = " % ", intercept_sep = " + "),
+    with_state = TRUE
   ))
 })
 
@@ -108,11 +152,15 @@ test_that("draw handles degenerate cases", {
     dts <- cut(pc$active_power, breaks = c(0, quantile(pc$active_power, probs = c(0.5, 1))))
     dts_cov <- data.frame(day_night = (pc$hour >= 7 & pc$hour <= 17))
     m_cov <- covlmc(dts, dts_cov, min_size = 10, keep_data = TRUE)
-    expect_snapshot_output(draw(m_cov, model = "coef", control = draw_control(time_sep = " % "), with_state = TRUE, digits = 2))
+    expect_snapshot_output(draw(m_cov,
+      model = "coef",
+      control = draw_control(time_sep = " % ", digits = 2),
+      with_state = TRUE
+    ))
     expect_snapshot_output(draw(m_cov,
       model = "full",
-      control = draw_control(time_sep = " % ", intercept_sep = " + "),
-      with_state = TRUE, digits = 2
+      control = draw_control(time_sep = " % ", intercept_sep = " + ", digits = 2),
+      with_state = TRUE
     ))
   }
 })