+++ tests/testthat/test-covlmc_trim.R
@@ -14,19 +14,34 @@ test_that("trimming removes only what it should remove", {
       expect_true(object.size(t_m_cov) < object.size(t_m_cov_model))
       ## identical contexts
       expect_identical(
-        contexts(m_cov, type = "data.frame", hsize = TRUE, model = "coef", frequency = "detailed", metrics = TRUE, merging = TRUE),
-        contexts(t_m_cov_model, type = "data.frame", hsize = TRUE, model = "coef", frequency = "detailed", metrics = TRUE, merging = TRUE)
+        contexts(m_cov,
+          type = "data.frame", hsize = TRUE, model = "coef",
+          frequency = "detailed", metrics = TRUE, merging = TRUE
+        ),
+        contexts(t_m_cov_model,
+          type = "data.frame", hsize = TRUE, model = "coef",
+          frequency = "detailed", metrics = TRUE, merging = TRUE
+        )
       )
       expect_identical(
-        contexts(m_cov, type = "data.frame", hsize = TRUE, model = "coef", frequency = "detailed", metrics = TRUE, merging = TRUE),
-        contexts(t_m_cov, type = "data.frame", hsize = TRUE, model = "coef", frequency = "detailed", metrics = TRUE, merging = TRUE)
+        contexts(m_cov,
+          type = "data.frame", hsize = TRUE, model = "coef",
+          frequency = "detailed", metrics = TRUE, merging = TRUE
+        ),
+        contexts(t_m_cov,
+          type = "data.frame", hsize = TRUE, model = "coef",
+          frequency = "detailed", metrics = TRUE, merging = TRUE
+        )
       )
       ## still usable with keep_mode==TRUE
       expect_equal(
         loglikelihood(m_cov, newdata = dts, newcov = dts_cov),
         loglikelihood(t_m_cov_model, newdata = dts, newcov = dts_cov)
       )
-      expect_no_error(contexts(t_m_cov_model, type = "data.frame", hsize = TRUE, model = "full"))
+      expect_no_error(contexts(t_m_cov_model,
+        type = "data.frame", hsize = TRUE,
+        model = "full"
+      ))
       expect_equal(
         simulate(m_cov, nsim = 50, seed = 0, dts_cov),
         simulate(t_m_cov_model, nsim = 50, seed = 0, dts_cov)
@@ -75,14 +90,27 @@ test_that("trimmed model can be drawn", {
       m_cov <- covlmc(dts, dts_cov, min_size = 10, keep_data = TRUE)
       t_m_cov_model <- trim(m_cov, keep_model = TRUE)
       t_m_cov <- trim(m_cov)
-      expect_snapshot(draw(t_m_cov_model, model = "coef", with_state = TRUE, digits = 2, p_value = TRUE))
-      expect_snapshot(draw(t_m_cov_model, model = "coef", with_state = TRUE, digits = 2))
-      expect_snapshot(draw(t_m_cov, model = "coef", with_state = TRUE, digits = 2, p_value = TRUE))
-      expect_snapshot(draw(t_m_cov_model, model = "full", control = draw_control(time_sep = " % "), with_state = TRUE, digits = 2))
+      expect_snapshot(draw(t_m_cov_model,
+        model = "coef", with_state = TRUE,
+        control = draw_control(digits = 2), p_value = TRUE
+      ))
+      expect_snapshot(draw(t_m_cov_model,
+        model = "coef", with_state = TRUE,
+        control = draw_control(digits = 2)
+      ))
+      expect_snapshot(draw(t_m_cov,
+        model = "coef", with_state = TRUE,
+        control = draw_control(digits = 2), p_value = TRUE
+      ))
+      expect_snapshot(draw(t_m_cov_model,
+        model = "full",
+        control = draw_control(time_sep = " % ", digits = 2),
+        with_state = TRUE
+      ))
       expect_snapshot(draw(t_m_cov,
         model = "full",
-        control = draw_control(time_sep = " % ", intercept_sep = " + "),
-        with_state = TRUE, digits = 2
+        control = draw_control(time_sep = " % ", intercept_sep = " + ", digits = 2),
+        with_state = TRUE
       ))
     }
   }